+++ CHANGELOG.md
@@ -7,11 +7,14 @@ All notable changes to this project will be documented in this file. For commit 
   **New Features**:
   - Create long-live api tokens to interact with API.
   - Helpful swagger page for API usage.
-  
+
   **Notes**:
     - Replaced backend http framework with go standard library
     - Right-click Context menu in single click mode can target the item that was right-clicked. (only on desktop)
-  
+
+  **Bugfixes**:
+  - Fixed ui bug with shares with password
+
 ## v0.2.10
 
   **New Features**: